chore: Add eventId on ErrorBoundary
