+++ src/config/localization/translations.json
@@ -1522,5 +1522,6 @@
   "here": "here",
   "For more details on the swap, please refer here.": "For more details on the swap, please refer here.",
   "Insufficient WBNB balance": "Insufficient WBNB balance",
-  "Select": "Select"
+  "Select": "Select",
+  "Error Tracking Id": "Error Tracking Id"
 }
